Add a Zapier integration for Fixie. (#28)
This is a Zapier integration that sends messages to Fixie when
triggered. It can be hooked up to anything that Zapier can communicate
with, e.g., Slack, Gmail, etc.

This is currently in the agents/zapier directory because I would like to
add a Zapier agent as well, but for now this is only the Zapier side.