Add agent for testing, and tests using this agent. (#38)
This PR adds a "tester agent" that uses the LLM to evaluate whether a
response from another agent meets certain criteria. This allows us to
automatically test the agents deployed through the fixie-examples repo.

This seems to work pretty well and gives us a lot of flexibility in
terms of how we express the conditions under which a given test should
pass. I realize a lot more work could go into this, but I think having
at least this kind of smoke test is useful to ensure that agents have
not flat-out broken after a push.